back to long long

+++ interface/Time.h
@@ -1,6 +1,7 @@
 #ifndef DBCommon_Time_h
 #define DBCommon_Time_h 
 namespace cond{  
-  typedef unsigned int Time_t;
+  //typedef unsigned int Time_t;
+  typedef unsigned long long Time_t;
 }
 #endif
